backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
greenlet==3.1.1
asyncpg==0.30.0
alembic==1.13.3
redis==5.1.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.2
lttb==0.3.2
scikit-learn==1.5.2
joblib==1.4.2
nltk==3.9.1
transformers==4.45.2
torch==2.4.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
